refactor(api): extract requirements linear program into own file
