+++ api/src/functions/query-requirements/domain/errors.ts
@@ -1,0 +1,16 @@
+const INVALID_ITEM_NAME_ERROR =
+    "Invalid item name provided, must be a non-empty string";
+const INVALID_WORKERS_ERROR =
+    "Invalid number of workers provided, must be a positive number";
+const UNKNOWN_ITEM_ERROR = "Unknown item provided";
+const TOOL_LEVEL_ERROR_PREFIX =
+    "Unable to create item with available tools, minimum tool is:";
+const INTERNAL_SERVER_ERROR = "Internal server error";
+
+export {
+    INVALID_ITEM_NAME_ERROR,
+    INVALID_WORKERS_ERROR,
+    UNKNOWN_ITEM_ERROR,
+    TOOL_LEVEL_ERROR_PREFIX,
+    INTERNAL_SERVER_ERROR,
+};
